feat: Add user-specific functionality to product selling lists
This commit brings enhancements to product-selling views of the app, enhancing the filtering of items according to user specificity. An 'id' parameter is now required for both 'getUserActiveSells' and 'getUserArchivedSells' functions in 'userData.ts' and optional parameters passed to 'ActiveSells' and 'ArchivedSells' in 'Profile.tsx'. This allows displaying user-specific active and archived sells.

On the server side, 'id' is extracted and used in database querying for 'active' and 'archived' sells in 'active.php' and 'archived.php' correspondingly. Additionally, some  redundant data retrievals and types are cleaned in 'ActiveSells.tsx' and 'ArchivedSells.tsx' for better readability and performance.